app/requirements.txt
fastapi==0.100.0
uvicorn
gunicorn
boto3
python-multipart
mangum
python-dotenv